Add contact page with form, refactor other files.
