+++ portfolio/src/main/webapp/favouriteAnimals.html
@@ -9,9 +9,9 @@
     <div class="navbar">
         <a href="index.html">Home</a>
         <a href="index.html">About Me</a>
-        <a class="current" href="projects.html">Projects</a>
-        <a href="favouriteAnimals.html">Favourite Animals</a>
-        <a href="index.html">Contact Me</a> 
+        <a  href="projects.html">Projects</a>
+        <a class="current" href="favouriteAnimals.html">Favourite Animals</a>
+        <a href="contact.html">Contact Me</a> 
     </div>
     <div id="content">
       <h1>Images Of My Favourite Animals</h1>
