+++ portfolio/src/main/webapp/index.html
@@ -12,7 +12,7 @@
         <a href="#about">About Me</a>
         <a href="projects.html">Projects</a>
         <a href="favouriteAnimals.html">Favourite Animals</a>
-        <a href="#contact">Contact Me</a> 
+        <a href="contact.html">Contact Me</a> 
     </div>
     <div id="content">
       <h1>Jahres Peter's Portfolio</h1>
@@ -35,7 +35,7 @@
         <div id="fact-container"></div>
       </div>
       <div id="contact">
-          <h2>Contact</h2>
+          <h2>Reach Out</h2>
           <ul>
               <li>
                 <img src="images/LinkedIn.png" class="contactIcons"/>:
