+++ portfolio/src/main/webapp/projects.html
@@ -12,7 +12,7 @@
         <a href="index.html">About Me</a>
         <a class="current" href="projects.html">Projects</a>
         <a href="favouriteAnimals.html">Favourite Animals</a>
-        <a href="index.html">Contact Me</a> 
+        <a href="contact.html">Contact Me</a> 
     </div>
     <div id="content">
       <h1>Jahres Peter's Projects</h1>
