+++ portfolio/src/main/webapp/script.js
@@ -31,43 +31,3 @@ function addRandomFact() {
     const factContainer = document.getElementById('fact-container');
     factContainer.innerText = fact;
   }
-
-
-/** 
- * SlideShow Functions
-*/
-
-  let slideIndex = 1;
-  showSlides(slideIndex);
-
-  function plusSlides(n) {
-      showSlides(slideIndex += n);
-  }
-
-  function curSlide(n) {
-      showSlides(slideIndex = n);
-  }
-
-  function showSlides(n) {
-    let i;
-    let slides = document.getElementsByClassName("slides");
-    let dots = document.getElementsByClassName("periods");
-    if (n > slides.length) {
-        slideIndex = 1;
-    }
-
-    if (n < 1) {
-        slideIndex = slides.length;
-    }
-
-    for (i = 0; i < slides.length; i++) {
-      slides[i].style.display = "none";
-    }
-
-    for (i = 0; i < dots.length; i++) {
-      dots[i].className = dots[i].className.replace(" active", "");
-    }
-    
-    slides[slideIndex-1].style.display = "block";
-    dots[slideIndex-1].className += " active";
-  }
